requirements.txt
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "3.11"
cffi==1.15.1 ; python_version >= "3.9" and python_version < "3.11"
charset-normalizer==2.1.1 ; python_version >= "3.9" and python_version < "3.11"
contourpy==1.0.6 ; python_version >= "3.9" and python_version < "3.11"
cycler==0.11.0 ; python_version >= "3.9" and python_version < "3.11"
deprecated==1.2.13 ; python_version >= "3.9" and python_version < "3.11"
fonttools==4.38.0 ; python_version >= "3.9" and python_version < "3.11"
googleapis-common-protos==1.57.0 ; python_version >= "3.9" and python_version < "3.11"
grpc-gateway-protoc-gen-openapiv2==0.1.0 ; python_version >= "3.9" and python_version < "3.11"
grpcio-tools==1.48.1 ; python_version >= "3.9" and python_version < "3.11"
grpcio==1.51.1 ; python_version >= "3.9" and python_version < "3.11"
idna==3.4 ; python_version >= "3.9" and python_version < "3.11"
inflection==0.5.1 ; python_version >= "3.9" and python_version < "3.11"
kiwisolver==1.4.4 ; python_version >= "3.9" and python_version < "3.11"
matplotlib==3.6.2 ; python_version >= "3.9" and python_version < "3.11"
numpy==1.24.0 ; python_version < "3.11" and python_version >= "3.9"
packaging==22.0 ; python_version >= "3.9" and python_version < "3.11"
pandas==1.5.2 ; python_version >= "3.9" and python_version < "3.11"
pillow==9.3.0 ; python_version >= "3.9" and python_version < "3.11"
protobuf==3.20.3 ; python_version >= "3.9" and python_version < "3.11"
pycparser==2.21 ; python_version >= "3.9" and python_version < "3.11"
pynacl==1.5.0 ; python_version >= "3.9" and python_version < "3.11"
pyparsing==3.0.9 ; python_version >= "3.9" and python_version < "3.11"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "3.11"
python-dotenv==0.21.0 ; python_version >= "3.9" and python_version < "3.11"
pytz==2022.7 ; python_version >= "3.9" and python_version < "3.11"
requests==2.28.1 ; python_version >= "3.9" and python_version < "3.11"
scipy==1.9.3 ; python_version >= "3.9" and python_version < "3.11"
setuptools-scm==7.1.0 ; python_version >= "3.9" and python_version < "3.11"
setuptools==65.6.3 ; python_version >= "3.9" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.9" and python_version < "3.11"
toml==0.10.2 ; python_version >= "3.9" and python_version < "3.11"
tomli==2.0.1 ; python_version >= "3.9" and python_version < "3.11"
typing-extensions==4.4.0 ; python_version >= "3.9" and python_version < "3.11"
urllib3==1.26.13 ; python_version >= "3.9" and python_version < "3.11"
wrapt==1.14.1 ; python_version >= "3.9" and python_version < "3.11"
